+++ const_panic_proc_macros/src/derive_debug/attribute_parsing.rs
@@ -17,6 +17,8 @@ use alloc::vec::Vec;
 use core::marker::PhantomData;
 
 mod keyword {
+    syn::custom_keyword!(display_fmt);
+    syn::custom_keyword!(panicvals_lower_bound);
     syn::custom_keyword!(debug_print);
     syn::custom_keyword!(ignore);
 }
@@ -31,6 +33,8 @@ pub(crate) enum ParseCtx<'a> {
 struct ParsedAttributes<'a> {
     debug_print: bool,
     crate_path: syn::Path,
+    display_fmt: Option<syn::Expr>,
+    panicvals_lower_bound: Option<syn::Expr>,
     impls: Vec<ImplHeader>,
     gen_params_props: Vec<GenParamProps<'a>>,
     type_const_params: Vec<Ident>,
@@ -40,6 +44,8 @@ struct ParsedAttributes<'a> {
 pub(super) struct Configuration<'a> {
     pub(super) debug_print: bool,
     pub(super) crate_path: syn::Path,
+    pub(super) display_fmt: Option<syn::Expr>,
+    pub(super) panicvals_lower_bound: Option<syn::Expr>,
     pub(super) impls: Vec<ImplHeader>,
     pub(super) gen_params_props: Vec<GenParamProps<'a>>,
     _marker: PhantomData<&'a ()>,
@@ -49,6 +55,8 @@ pub(super) fn parse_attributes<'a>(ds: &'a DataStructure<'a>) -> syn::Result<Con
     let mut this = ParsedAttributes {
         debug_print: false,
         crate_path: syn::parse_quote!(::const_panic),
+        display_fmt: None,
+        panicvals_lower_bound: None,
         impls: Vec::new(),
         gen_params_props: ds
             .generics
@@ -139,6 +147,16 @@ fn parse_helper_attribute<'a>(
         check_is_container(&ctx, empty)?;
 
         this.debug_print = true;
+    } else if let Some(_) = input.peek_parse(keyword::display_fmt)? {
+        check_is_container(&ctx, empty)?;
+
+        input.parse::<Token!(=)>()?;
+        this.display_fmt = Some(syn::Expr::Verbatim(input.parse()?));
+    } else if let Some(_) = input.peek_parse(keyword::panicvals_lower_bound)? {
+        check_is_container(&ctx, empty)?;
+
+        input.parse::<Token!(=)>()?;
+        this.panicvals_lower_bound = Some(syn::Expr::Verbatim(input.parse()?));
     } else if let Some(_) = input.peek_parse(Token!(crate))? {
         check_is_container(&ctx, empty)?;
 
@@ -200,6 +218,8 @@ fn finish<'a>(
     let ParsedAttributes {
         debug_print,
         crate_path,
+        display_fmt,
+        panicvals_lower_bound,
         impls,
         gen_params_props,
         type_const_params: _,
@@ -209,6 +229,8 @@ fn finish<'a>(
     Ok(Configuration {
         debug_print,
         crate_path,
+        display_fmt,
+        panicvals_lower_bound,
         impls,
         gen_params_props,
         _marker,