+++ Changelog.md
@@ -2,6 +2,15 @@ This changelog is a summary of the changes made in each release.
 
 # 0.2
 
+### 0.2.12
+
+Added these helper attributes to `PanicFmt` derive and `impl_panicfmt` macro:
+- `#[pfmt(display_fmt = ...)]`
+- `#[pfmt(panicvals_lower_bound = ...)]`
+
+Added `const_panic::utils::{min_usize, max_usize}`
+
+
 ### 0.2.11
 
 Added formatting for `core::num::{IntErrorKind, ParseIntError}`